python-reference/requirements.txt
fastapi
uvicorn
supabase
google-api-python-client
google-auth
python-dotenv
pydantic
